requirements.txt
dotenv
asgiref
mysqlclient
celery
channels[daphne]
django
django-allauth
django-cleanup
websocket
channels_redis
google_api_python_client
ratelimit
luscioustwitch
markdown
pillow
requests
